requirements.txt
requests~=2.28.2
mcstatus~=10.0.1
uvicorn~=0.20.0
fastapi~=0.89.1
srvlookup~=3.0.0
pillow~=9.4.0
aiofiles~=22.1.0